Final Tutorial Step 1 Complete
At this point, the wishlist is obtained from the database.  Changes included the addition of the immutability-helper as well as treating the wishlist as an object with an array of products.

Biggest issue was figuring out promises - putting the HTTP request in wishlist.js was easy, making the dataService handle the data was more tricky.