Intregrated React Router 4
Finally figured out how to use React Router to change between "pages" for a  single page website.  I may leave this project as is, since the next step of actually implementing a button or whatever to create a wishlist, manage wishlists, etc. is an exercise that has been done in other forms here.